El de la mañana 22/07/2019
Se crea over y estado de resultados

+++ backend/database/migrations/2019_07_05_145918_create_ahorro_table.php
@@ -22,8 +22,8 @@ class CreateAhorroTable extends Migration
             $table->integer('personal_id')->nullable();
             $table->date('fechapago')->nullable();
             $table->string('nombre')->nullable();
-            $table->integer('montopagado')->nullable();
-            $table->integer('limiteahorro')->nullable();
+            $table->decimal('montopagado',10,2)->nullable();
+            $table->decimal('limiteahorro',10,2)->nullable();
             $table->string('status')->nullable();
             $table->timestamps();
         });
